added: support for specifying initial conditions from g2 files
use <initialcondition field="fieldname" step="x" comp="y"/>
where
 - fieldname is integrand specific
 - step is the time level
 - comp is the component (for vector fields)

currently it assumes that the given g2 file has the same basis
as the geometry, and simply extracts the third component of the
control point coefficients as the data. this means that we are
currently limited to 2D, and that we cannot handle vector fields.

this needs to be revisited for generality.

git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@1924 e10b68d5-8a6e-419e-a041-bce267b0401d