fixed: replace the 'real' define with 'Real'
'real' is a dangerous symbol to reserve in the global namespace,
and this has been a ticking bomb all the time. it finally blew,
if gotools is compiled against boost, it will pull the <complex>
header and the define clashes with the real() member.

while still a define in the global namespace, Real should be less
class prone

git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@1947 e10b68d5-8a6e-419e-a041-bce267b0401d