fixed: load dependent fields in the HDF5 output routine
may be needed for derived quantites (secondary solutions).

note that this is order sensitive, in particular we want to do the
call that loads only the solution vector last. this is needed as
certain SIMs (such as K-Epsilon) has a dependency to themself to be able
to handle the complex boundary conditions. due to this, if we call the
dependency loading function last, we end up with the wrong vector loaded.

git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@2302 e10b68d5-8a6e-419e-a041-bce267b0401d