Fixed: Handling of chained multi-point constraints.
For problems with time-varying dirichlet conditions,
the constraint equations were not properly updated
between the increments. This was due to the resolving
where we lost the connection between the actual master dof
beeing updated, and the slave in the other end of the chain.

The update is now done in a recursive manner instead, using
a pointer to the next MPC in the chain stored on the DOF level.

Notice that this fix also affected the FSI-slip regression test.

git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@2354 e10b68d5-8a6e-419e-a041-bce267b0401d