Added: Method to inject patch vectors using a specified madof array.
Allows to specify a basis to inject for. This will be used for mixed models
where we have DOF vectors with additional components on the first basis.

Also renamed the associated extractNodeVec method to avoid ambiguity, and
made the error handling of this method equivalent to the new inject method.