fixed: due to the order of things, we need a model generator member
due to the structure of parsing there is no way around this.
we first need to generate the model (and topology sets) in the SIMbase
parse function. we then need to refine the model in the SIMxD parse
functions which are called after the SIMbase parsing. thus,
we need to hand the model generator instanced in the SIMbase parsing
to the SIMxD to create the topology at the appropriate time.