changed: check if an empty secondary solution is associated with residual norms
if so, count the norm entries.

also add a flag for externally provided projections in the norm class.
this is necessary for applications where the projection is performed
on a different basis than than the first finite element basis. in particular,
this is used with subgrid and taylor-hood discretizations of stokes