Added: virtual method evalSol2 and non-virtual IntegrandBase::evalSol1.
The latter method can be used when the secondary solution depends on
the first primary solution vector only (which covers most cases).
Then the integrand can override the evalSol2 method instead of evalSol
such that they don't need to deal with the element-level extraction
of solution vectors themselves (the default evalSol will do it).
Also extended the doxy for the initElement and evalSol methods.