Fixed: Incorrect sanity check correction.
It turns out I was wrong in e3da110cb9ea7f4fa3634257f64abcdf531782eb.
The check should instead use the number of coefficient in the patch,
which will work both for mixed and regular problems.
The check against MLGN.size() is still needed as a secondary condition,
in case of dirichlet conditions in local axes (with extra nodes around).