IFEM-test: suppress listing on non-root processes
Parallel unit tests were registered once per rank they run on in ctest,
as all processes output the list when run with --gtest_list_tests.
This is the mechanism used to register the tests in ctest. By suppressing
listing on non-root ranks, we now only get a single register as intended.